+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ## 2.21 branch
 ### 2.21.5
 * added `iframe()` and `youtube()` methods to the `BootstrapHtmlHelper`;
+* the `HtmlHelper` is now totally deprecated, use `BootstrapHtmlHelper` instead. In a later version, the latter
+  will take the place of the former, assuming its name;
 * the `DropdownHelper` is now totally deprecated, use `BootstrapDropdownHelper` instead. In a later version, the latter
   will take the place of the former, assuming its name;
 * the `FormHelper` is now totally deprecated, use `BootstrapFormHelper` instead. In a later version, the latter will